can add shop items to depo, need fix dialog showup, delete added items

+++ core/domain/src/main/java/hu/tb/core/domain/product/Product.kt
@@ -2,6 +2,6 @@ package hu.tb.core.domain.product
 
 data class Product(
     val name: String,
-    val quantity: Int,
+    val quantity: Double,
     val measure: Measure
 )
